pyproject.toml
[tool.poetry]
name = "mlflow-workflow-app"
version = "0.1.0"
description = ""
authors = ["Heiko Kulinna <Heikokulinna@gmx.de>"]
readme = "README.md"
packages = [{include = "mlflow_workflow_app"}]

[tool.poetry.dependencies]
python = ">3.8.1,<3.9"

mlflow = "^2.4.1"
pandas = "^2.0.1"
numpy = "^1.24.3"

dash = "^2.9.3"
flask="*"
dash-extensions = "^1.0.0"
dash-core-components = "^2.0.0"
dash-html-components = "^2.0.0"
dash-carbon-components = "^0.4.2"
dash-daq = "^0.5.0"
python-dotenv = "^1.0.0"


azure-storage-blob="*"   # 12.16.0
azure-identity="*"  # 1.13.0

plotly = "^5.14.1"
sklearn="*"    #0.0.post5
PyYAML="*"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
ruff = "^0.0.267"
pre-commit = "^3.3.2"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79


